Fix the ugly case of reference to dynamic array.
The size of this array comes from the network.
It is needed both at the reference reception level (to allocate enough
space) and at the array reception level (to fill enough room).

It is then passed as an argument of the recursive function. This is a crude
hack, but I was told that working code is sometimes better than neat one ;)


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@147 48e7efb5-ca39-0410-a469-dd3cf9ba447f