Most of this change is the consequence of the refactoring of the name of the function connected to the mutex object and of the unsupported semaphore function on MAC OS X.
The file synchro.h rename the functions xbt_mutex_acquire() and xbt_mutex_release() to xbt_mutex_acquire() an respectively xbt_mutex_release().
This file also contains the declaration of two new functions xbt_mutex_tryacquire() and xbt_mutex_timedacquire(). The first
is the non blocking version of the function xbt_mutex_acquire() (if the lock is already owned by an other thread the function immediately
returns). The second function xbt_mutex_timedacquire() attempts to get the mutex during a specified delay (timeout), if the lock is not available
during this delay the function returns (throws a timeout exception). These functions (for the real life) are defined in the file xbt_rl_synchro.c and (for the simulation) in the file xbt_sg_synchro.c.

The header xbt_os_thread.h declare two new functions xbt_os_mutex_tryacquire() and xbt_os_mutex_timedaquire(). These functions represent
the implementation of the xbt_mutex_tryacquire() and xbt_mutex_timeacquire() for the real live.
These functions are implemented in the source code file xbt_os_thread.c.

The change in the files dynar.c and xbt_queue.c, philosopher.c and parallel_log_crashtest.c is the consequence of the refactoring of the functions previously described.






git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@4866 48e7efb5-ca39-0410-a469-dd3cf9ba447f