SIMIX_network_copy_data() is a mess because it tries to allow passing arbitrary long memory chuncks in simix_network. It complicates stuff, and is hopeless anyway since it has not the full gras_datadesc power. So, let's stop trying that.
When SIMIX_network_send(), don't pass a pointer's address (void**) as
argument src_buf (which is declared void*). It was dereferenced twice
in SIMIX_network_copy_data() anyway.

The current situation is a bit more sane, even if further cleanups is
needed: I guess that the data argument of SIMIX_network_send() become
useless now, and I didn't cleanup my commit: that's full of useless
FIXMEs.

The motivation to this change was that ruby stacks may move to some
other location in the memory on context switch (to extend it). So,
trying to reuse a pointer to a variable living in the stack after a
switch does not work. This commit kills a "invalid read".

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@7199 48e7efb5-ca39-0410-a469-dd3cf9ba447f