Improve the integration of mmalloc and mc_memory into the mess.
By default (ie, when passed a NULL mmalloc descriptor), mmalloc and
friends use a mmap based implementation (instead of the old, slow sbrk
one).

mc_memory makes sure that we pass something else than NULL as mmalloc
descriptor to mmalloc function only if the model-checker is used.

(the model-checker is not activable yet -- the integration is maybe
next commit)

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@7710 48e7efb5-ca39-0410-a469-dd3cf9ba447f