Call surf_parse_models_setup() from the right level
This function is in charge of initializing the models after all
configs have been parsed. It used to be called in the XML parser
callback, it is now called from the platform creation API, which is
intended to be called either from the XML parser or from the lua
console, or directly from the user code.