Avoid (almost) duplicated code in xbt_log_layout_*.
Remove the dynamic log layout function.

The layout functions now try to build the log string in a buffer of given
length, and return success or failure.

The first try is with a buffer of fixed size.
In case of failure, the buffer is expanded, until it is large enough.

With this change, the code is much simpler, and the only penalty is with
very large log strings.