Reimplement the lua console on top of sg_platf
I admit it was too early to do that change since sg_platf is not quite
fully separated from the XML parsing yet, but the cruft accumulates
from everywhere, so I have to break stuff a bit to fix the overall
picture. Yeah, the lua tests are not fixed yet.