Remove the complexity needed to save a global that is never used
this is the field "model" in the surf_model_description structure. The
actual model is always saved elsewhere (like in the global
surf_workstation_model and friends).

If one day we want to have multi-model simulations, we'll need to
partially revert this change, but I guess that this day seaching the
right model linearly by name, as it was made possible with the
existing setup, will not be the way to go.