Change again the calling convention of routing_get_route_and_latency
now, you cannot call it without passing a valid route receiver. It
will be created if pointing to NULL, or simply filled. Actually that
was the de facto calling convention in the code.

This allows to simplify routing_get_route_and_latency by not checking
if the caller is interested in the route at every line.

(plus some useless cosmetics)