Improve MSG doc a bit
It is impossible to document lua modules with doxygen so far, so don't
do a submenu to separate the C interface from the lua one, but add a
word about lua in a unique page, under MSG.

The C interface moves up directly under MSG, without any MSG_C in the
middle.

Move all MSG module creations into module-MSG.doc for simplicity (and
because this is the only way to finely control their order).

Also, drop channel documentation.