Optimize xbt_str_diff further.
Make it faster and less memory hungry when there are many lines and few
differences, for example when test tracing-trace_platform is failing...

The general idea is to reduce the problem size for diff_build_LCS/diff_build_diff.
This is done by finding, when it is easy, the start and the end of the result.
It's easy when the line are the same, or when a line from the one set does not appear
in the second set.