Stop using SIMGRID_ROOT at compile time, use SIMGRID_LIB/INCLUDE instead
At configuration time, the SIMGRID_ROOT variable is used as a hint in
the FindSimgrid chunk, when looking for our headers and libraries. If
they cannot be located anyway, the user is given the opportunity to
manually give the mandated pathes in ccmake for example.

Then, at compilation time (ie, once the configuration is done), we
should not use SIMGRID_ROOT anymore, but only SIMGRID_LIB and
SIMGRID_INCLUDE because otherwise, we ignore specific settings that
the user may have provided in ccmake.