Hide pretty nasty issue.
In this example, a process is killed right at the same time as it starts
executing somthing. It creates the surf action but is killed before it
gets the opportunity to wait on it. Consequently, the kill does not clean
up the surf action (so technically, the corresponding resources are still
wasted).

This could be "solved" by doing the kill before the simcall that created
the surf action. But more generally, it raises the issue of all the
data that have been allocated during the MSG_task_execute. There is
something really annoying with the kill. The current MSG implementation
is clearly not protected against kills. In case of kill, when doing a
simcall, we should throw an exception and TRY CATCH everywhere. We're far
from this for now so I leave it for another time.