Process that suspend themselves now create the corresponding surf action
A process that suspends himself did not create any surf action. When
processing the simcall, the maestro only set the process as suspended
and did not answer the simcall. Likewise a resume would simply wake up
the process.  A problem with this is that if the corresponding host ever
went down, the suspended process would never have realized.

For the moment, just die whenever a process is suspended and it is not
waiting on an action that can be be suspended. The consequence is that
there is no reason to wake up a process that did not have any waiting
action.  Again, DIE if this ever happens for the moment.