Add a new type of typed task that we may be able to autoschedule.
SD_TASK_COMP_PAR_AMDAHL represents a parallel tasks whose initial work
(amount parameter upon creation) is distributed among host according
to the Amdahl's law. Such tasks are created with a parameter alpha
that corresponds to the non-parallelizable part of the computation.

Before calling SD_task_schedulel or SD_task_schedulev, the
SD_task_distribute_comp_amdahl function has to be called with the
desired number of workstations. The filling of computation_amount and
communication_amount (actually empty) structures is then done seamlessly.

Not sure whether SD_task_distribute_comp_amdahl should be
automatically called within SD_task_schedulev or not. Easy to modify
if we want to hide everything. An option could to call it if the user
didn't do it.

move all the creation functions at the same place in the process.