Add some support to bench the sequential and parallel parts of a simulation
This is activated by defining the TIME_BENCH_AMDAHL macro at compile
time (for example by decomenting its definition in smx_private.h).

This is useful to evaluate the maximal possible speedup of a given
simulation by applying the Amdahl Law.

This could be changed into a command line flag at some point, but we
first have to assess the performance impact of playing that much with
the timers (should be limited, but who knows?).

The initial patch was from Jean-Noel Quintin, thanks for that!