Bug fix in mrealloc where the previous content could get lost
When reallocating a large block, mrealloc tries to be cleaver and
first frees the area before requesting it again to avoid a copy if
there is some free adjacent memory.

This trick obviously don't mix well with the mmalloc memseting the
allocated memory to 0 before returning it.

So, split the mmalloc in two, and use the non-memsetting one from
mrealloc.