Fix the parallel execution for the bittorrent example.
The problem was that RngStream creation is not thread safe.  Furthermore,
each created RngStream is dependent of the previously created RngStreams.

The problem is circumvented by creating the RngStreams before the launch
of the application.  Each host has its own stream, that can be retrieved
through MSG_host_get_data().