refine code around model types and objects
The model_type member of an action object is renamed model_obj. This
change makes it clear that this member points to a surf model objet, not
the type of a model.

The model type member is added to a surf model object. This allows us to
know what type a model object is and what extension field is accessible.