fix the segfault caused by null smx_action
A surf VM object has the dummy action that wraps all the execution tasks
on the VM. This dummy action has its object only in the surf layer and
not in the SIMIX layer. There may be a better solution. But, at now we
simply skip touching the null action object at the SIMIX layer.