Fix miscalculation of CPU shares on a multicore PM.
The recent commit adding the load capping feature broke CPU share
calculation on a multicore PM. This patch fixes the problem.

The bound value "0" means that there is no CPU bound for that task or
VM. On a single-core PM, this is true. The task can use all CPU resource
on that PM. On a multicore PM, the task can use the resource of
only a CPU core.

Even MSG_{task/vm}_set_bound() is not used in a simulation program,
simulation results will be wrong for multicore PMs. This is because
the bound value "0" is used internally.