Use more explicit variable names and add a bit of debug.
There is still something "wrong" with this code, which is that it aborts whenever an action gets a share of 0 (in which case I am not able to assign it a completion date and insert it into the calendar). This could happen but is more likely to be a bug in lmm or in surf, which is why I leave it as such for the moment.