[smpi] Allow zero-sized custom datatypes with MPI_Type_contiguous()
MPI_Type_contiguous(0, oldtype, newtype) is valid in MPI but SMPI does
not accept it as the resulting datatype is considered invalid: SMPI
does not accept datatypes with size of 0. Fix the datatype validation
function in order to accept this.

This type of datatype is used in bcastzerotype.

bcast-NTSL is fixed as well (division by zero).