[mc] Add a (compile time) option for content addressage page store
With a simple hash, we expect collision to happen: when inserting a
new page in the store, each stored page with the same hash is compared
(memcmp-ed) with the current page.  If a page already present is
inserted, the content of the stored page is always pulled in the
cache: when inserting n pages, into the page store at least 2×n pages
are pulled in the cache.

With this option, a collision-free fingerprint of each page is
computed.  As there is no collision, we do not have to check the
content of the stored pages.  MD4 is used as we do not need security
against malicious users and it is much faster than more sescure
cryptographic hashes.