Make it much easier to follow control flow through simix
One of the main difficulty in navigating the source code when debuging
was to find your way from MSG to SURF, through SIMIX. The simcall
mechanism really breaks the flow.

The solution is to add a dummy call to the simcall body, such as:
    if (0) SIMIX_pre_host_off(&self->simcall, host);

With the "if (0)", it will be cut at compilation and have no impact on
the code behavior. But you can go to that function in eclipse, press
F3 and you are within the kernel now. No need to grep your way around.