[smpi+mc] When restoring a snapshot, always remap the SMPI privatisation segment
After restoring the global variables in a snapshot, the global
variable containing the currently active SMPI privatization segment is
overwritten. As a consequence, the segment SMPI thinks active might be
inconsistent with the currently active segment (which the kernels
knows about).

In this case, we want to remap the segment regardless of what SMPI
thinks.

We might be more clever to avoid unecessary segment switch if the
segment which is really mapped is the same as the one we want.