[mc] Basic infrastructure for a real model-checker process
The model checker process communicates with the model-checked
application using socket (and wait). Currently it waits for the MCed
process initialisation and fetch its system state, DWARF information,
etc but does not do anything else.

The previous (standalone) mode is currently used by default. The new
behaviour is triggered with the SIMGRID_MC_MODE=server
environment. The idea is to keep the standalone version at least as
long as the new version is not stable/working.