Prevent warning message 'config elem ... registered twice'
Currently, Lua platform files require the MSG module to be
initialised. Hence, when SMPI and Lua platform files are
used at the same time, MSG and SMPI will be
completely initialised, resulting in multiple calls
to TRACE_global_init which sets up the config params.

This commit simply adds a flag to that function so that
the config elems are registered only once.