Hopefully fix the strdup/_strdup issue
The problem was related to the fact that, when compiling in C++11
standard mode (--std=c++11), the compiler/library would not let us use
the (non-C++11-standard) strdup()/_strdup() functions. However, the C
code was not affected (and the cmake function detection function was
not affected as well and was detecting the functions as available). We
have to compile in C++11-with-extensions mode (--std=gnu++11) in order
to get them.