[mc] Do not attempt to make the heap writeable
We add a mprotect()-based gap after the stacks (which are created on
the heap) in order to detect stack overflows but when the stack is
destroyed, the protection is removed by making the pages (part of the
heap) RW and executable. This is not a good idea :)

Apparently, it fails on Ubuntu because of PaX restrictions on
mprotect(): the mprotect calls fails, the pages are still unreadable
and unwritable and the application segfaults when a new allocation
attemps to use it.