S4U: Implement Storage and File
The interface is maybe a bit improvable, and the implementation
remains as twisted as it was, but at least it's usable from S4U.

Improving the implementation would mean moving the book keeping of
files and sizes out of surf. That could be a task for simix if it's
not at user level, but that's not really for surf. Actually, most of
the file API should be moved elsewhere. But I'm not sure of where.

Improving the interface would mean making explicit the storage on
which I create a file instead of relying on the prefix naming.