[MC] ensure that the main ISP/Umpire tests are activated
All 74 tests of http://formalverification.cs.utah.edu/ISP_Tests/ are
activated. The many other tests of the directory that were not listed
on that page are not activated.

The tesh files are generated on the fly as we don't check the output
anyway. cmake coding is such a pleasure, I could not resist.

We should probably add a way to use tesh without tesh files but only
with command line flags. But not tonight.

      /\                                        /\
     /!!\   These tests are currently broken   /!!\
    /_!!_\                                    /_!!_\