Simplify how shareResources() is called in each models
- Host->shareResources() is in charge of calling the ones of Cpu, Net
  and Storage
- VM->shareResources() must be called after the Host one

So, don't make a list `model_list_invoke` of two elements that must be
called in a very specific order. Instead, directly call these methodes
in surf_solve().

A next step could be to put the content of Host->shareResources and
maybe VM->shareResources() in surf_solve so that Host and VM can stop
being resources just to dispatch calls to the real resources.

But L07 must be dealed with before (eg by moving the content of
Host07->shareResource() into Cpu07->shareResources() and leaving
Link07->shareResources() empty).