Simplify the code taking Cpu::m_speedPeak changes into account
The DVFS change set failed to do the right thing and instead added the
current speed as a bound to any execution action (in addition, this
code was in simix while this performance modeling clearly belongs to
surf). I guess that this was not playing well with availability
changes due to a trace, since that extra bound was not recomputed
correctly.

Instead, factorize the code of trace availability events with the
m_speedPeak changes created by DVFS. This creates a new Cpu event
called onSpeedChange, which is still to be exported properly. For now,
only the tracing code listen to that event (it's hardcoded).