Fix broken build 20551fb51e6144ee0c7942602e4c4a60477ddfc8
This commit fixes compatibility issues with GCC < 5 and older versions of
clang.

The problem (at least with GCC) is/was that they declared the move-constructor
as deleted in GCC < 5. See the bug: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=54316

This commit does not return any stringstreams any more but makes use of a
file-wide global variable to use stringstream.