[SMPI] Introduce variable computation speeds.
This commit introduces the possibility to use a csv file that
contains sections of the code that should be sped up or slowed down
according to the second column in that specific file.

One line could look like this:
exchange_1.f:30:exchange_1.f:130,1.18244559422142

This means that computation time between line numbers 30 and 130 offile exchange_1.f
should be slowed down by a factor of 1.18.

Of course, this only makes sense if computation time is calculated.

Activation via the newly introduced smpi/comp-adjustment-file CLI cfg param:
--cfg=smpi/comp-adjustment-file:/path/to/file.csv