so much dupplicated code, this is disgusting.
The existing code was dealing specifically with the first matching
action to initialize the minimum. I just update the value
inconditionnaly when it's still at its initial value.

We should use numeric_limits<double>::max() instead of -1 as an
initial value, but we return it when no action can be found, and I
don't feel like fixing the expectations of all calling code here.