[ENERGY] Introduce idle/1core/maxCore energy consumption
Up until now, we only supported idle and max_core consumption.
We then used a linear model to compute the energy that we used.

This is not an accurate model, since the difference between idle and 1 core
and 1 core and 2 cores is not equal. We hence need the first core.

But even that is not 100% accurate; for the moment it will have to suffice, though.