fix gcc 6.2 builds
It was failing with the error: type 'struct yy_buffer_state'
violates the C++ One Definition Rule.  A different type is defined in
another translation unit.

This comes from the fact that I rebuilt one of the flex parser with a
new version of flex. Rebuilding all of them is difficult because they
are somehow manually edited to make the C compile with g++.
Dirty, dirty :(