[SMPI] Use SmpiHost::orecv().
I moved the computation of the orecv configs to the SmpiHost class.
It's currently a crude implementation and does not (yet) allow for a
per-host configuration, but this is planned and outlined in the code.
For that, all we need is an option to parse and set orecv values on
a per-node basis. (See the orecv_ property.)