fix a memleak in Java VM when the used did not manually destroy VM
The C world was cleaned on a vm.destroy() and not when the GC
finalizes the Java world.

- Pb1: any java call to a destroy()ed vm could segfault
- Pb2: C world memleaked when forgetting destroy()

So now, the VM is not really destroyed when calling destroy(), but
only when all references to the object are gone. It tastes much more
Java to me that the previous version where our users had to to do the
memory management manually.

Arguably, this differs a bit more from libvirt, but I think I can live
with it.