SMPI: simplify process creation
finalization_barrier is either created when the processes are created
from smpirun, or taken from the instance if any. So there is no need
for a setter

This could be further simplified by making a default instance when
starting from smpirun. I just need to understand how to make this move
in the current swamp of corner cases...