Portability fun with FreeBSD
Make sure that the lambda capture parameter is actually used when
sendfile(2) is not usable, to avoid a fatal warning: FreeBSD is both
super advanced (with clang being very uptodate) and somewhat
prehistorical.

And no, I do not plan to deal myself with the FreeBSD-specific
sendfile prototype to avoid this problem in a cleaner way. But patches
are welcome as usual.