Allow multicore VMs, and take this into account in the sharing
This does not break any test, because it's not well tested yet. Yup,
the ChangeLog is a bit overselling so far. A new test torturing this
to make sure we get the right result in many sharing situation is
still to be written. examples/msg/cloud-simple will be taken as a
basis for that.