further WIP on the ActivityImplPtr feature. Now it compiles (but fails)
ActivityImplPtr are all over the place, and the manual refcounting is
dead. Most tests still fail, for some reason that I don't understand
yet. That's a WIP...

Notable point:
MC wants to retrieve the pointers to the ActivityImpl in a remote
process, even if their type is ActivityImplPtr (that is,
intrusive_ptr<ActivityImpl>).

The good thing is that in this case, the raw pointer is stored in the
simcall parameters, and the *Ptr type is reconstructed by the marshal
function.

So I just have to generate the <simcall>__getraw__<param/res>()
functions that return the raw pointer without building a *Ptr type. It
compiles, but the MC is broken ATM so maybe it's not sufficient.

I'm committing anyway because I want to add simpler tests for the
communications that would make the debug of this feature easier to
achieve.