fix 32-bit storage tests
The modification of the N11 storage model uses a rounding method so
that entire numbers of bytes are written at each action update. This
causes a slight difference of 30 nanoseconds in the total execution
time of the example. Assume that we can leave with that and just print
time up to the millisecond. Moreover this model is going to change
again soon.