Fix smpi_execute_benched.
When simulating a target machine T on an host H, you had to give the
time as measured on H. This time was then “converted” to the time that
would be spent on T by using the host speed and the target speed.

Now, you directly give the time that should be spent on T. This is
better, since the relative duration of different pieces of code may differ
from H to T (and thus you really want to give the time spent on T).