getting rid of bprintf raises a potential bug :-/
had to add a grace time on the master after sending the 'finalize'
order to all workers. Without it, master and Bourassa end at the same
time (last worker to be stopped). However, it seems the master
finishes first and let bourassa in a 'running' state :
Process 6 (worker@Bourassa): waiting for communication synchro 0x24bd6e0 () in state 0 to finish

Commenting line 54 triggers the issue.