Disable broken test.
It gives a yellow bullet on Jenkins, and doesn't bring anything.

More precisely, the test aborts early because MPI_Type_extent doesn't give
the expected result (Unexpected extent for small struct). The result of
the test was however never verified, since the test is only used by the
model checker.

On the FreeBSD node (clang/libc++), something is broken with exception
handling and the model checker hangs indefinitely at context's ::stop(),
when a StopRequest exception is launched.

The test can be reactivated either when MPI_Type_extent is fixed, or an
update corrects the problems on FreeBSD.