[SMPI] Privatization changes, step 1
This commit introduces a new variable that retains the whole initial
data segment during the execution of the program. We need this
as we will soon allow new processes to be added dynamically and we
can hence no longer assume that we can initialize all processes at the
beginning of the execution.