nobody will launch the logging AFTER the application start
So there is no need to take a timestamp when the loggin is started,
and then substract it from the surf clock.

Plus, calling MSG_get_clock() from xbt_log_layout_simple_new() is not
a good idea since the later is called from xbt_log_preinit(), which
occures before MSG_get_clock() is usable (not all MC stuff is inited
on centos).