Reduce code duplication.
The only difference between Vector and Hvector, and between Indexed and
Hindexed is that stride and block_indices are given in multiples of
old_type->get_extent() in the former case, while it's given in bytes in the
latter case.
Thus, make Vector a subclass of Hvector, and Indexed a subclass of Hindexed.