for some reasons, the PID was sometimes not inited correctly in Java
It is supposed to be inited during actor creation, but I've seen
during the SUD'17 a machine where it remained at -1. So, if the PID is
negative, java will ask again to the C for a fixup. We could have
fixed the problem, but chasing down JNI issues is really hard.

I prefer to play nasty and safe here. Or let's call it the belt and
brace approach instead.