Don't test features that are impossible to get right
There is no automatic refcounting in C so when a process is killed, we
must free it immediately. We could maybe propose a workaround to allow
the users to increase the refcount themself (as needed for #248), but
it's much easier to convert your code to C++ than getting the manual
refcounting right, I'd say. Sorry...

The fix to #248 leads to invalid accesses detected by valgrind when
applied to C/MSG.