[SMPI] Next step for dynamic privatization
This commit breaks the following tests:
The following tests FAILED:
        181 - smpi-msg-masterslave-thread (Failed)
        182 - smpi-msg-masterslave-ucontext (Failed)
        183 - smpi-msg-masterslave-raw (Failed)

To fix this, many more changes will be required. This will be done in
the next commit. This commit is just there to make sure changes
remain easily understandable (by having small, easy-to-read commits).

I changed the following:

- Remove variable index_to_process_data
- Make process_data a map instead of an array.
- Remove the "index" variable for SMPI instances (no longer needed)

And cleanup the mess that comes with these changes (delete/free, ...)