Un-deprecate MSG_task_get_flops_amount() and implement it correctly
The problem that pushed us into deprecating it is that for parallel tasks,
there is no way that this function can return an amount of flops:
the remaining work is a vector, not a scalar.

So for parallel tasks, you can only get a completion ratio.

Our fix was to rename this into MSG_task_get_remaining_work_ratio(),
but that's ... suboptimal because for sequential tasks, this actually
returns a scalar (the amount of flops) and not a ratio...

So this commit is the first one in a serie of three:
- here, we implement the correct MSG_task_get_flops_amount()
- next we will stop using MSG_task_get_remaining_work_ratio() where we
  expected flops and only keep it where we expected a ratio (ie, in
  the recently added example)
- next, we will make sure that MSG_task_get_remaining_work_ratio()
  always returns a completion ratio, even for sequential tasks.
- next, we can remove the MSG_task_get_initial_flops_amount() that was
  added as a first attempt to fix this mess

Stay tuned :)