Pliz don't have two log channels with the same name
This was breaking in a weird way. On Linux, the internal replay
channel was used for both the parameter setting and the use (it
resulted in more things to be shown when setting replay.thres:verb).
On Mac, the parameter setting did not reach the channel usage, and
nothing was shown.

The fact that the linker accepted these two competing definitions of
strong identifiers is weird. According to my teachings, this should
have failed :)