fix #245: migrating an actor does not migrate its execution
Before, the migration was only taking place when the actor got awaken
after an activity. Now, it takes place right away, and if it's
blocking on an execution, the activity is also migrated. If it's another
kind of activity, then an error is raised as this is not implemented
yet.

Also rewrite the s4u-actor-migrate example to be less funny but more
informative, and to test that feature.

(fix #245)