Messing up with VM  - move VM tracing to S4U  - move last functions related to migration to the plugin     - MSG_vm_create is now MSG_vm_create_migratable and tagged       as DEPRECATED (will be removed in 3.21)     - Assume that the live migration plugin is always loaded in JAVA       (because all VMs are create with MSG_vm_create there)  - add a proper destroy function to S4U VMs (shutdown first, then    destroy)  - pimpl_vm_ is now private (with getImpl() as accessor)     - and use the user level interface as much as possible
Conclusion: The MSG_vm interface is now just a dummy wrapper on the
S4U interface and live migration is fully isolated in a plugin. Only
the is_migrating bool remains in VirtualMachineImpl, but is not (and
should not be) accessed or modified from outside the plugin. We can live
with that.