[SMPI] Remove index notion from smpi::Group
This is a huge commit and removes the index() function from
smpi::Group and replaces it with an actor() function.

The index was basically just Actor::getPid()-1 and doesn't need
to be saved. It was also often used for the privatization segment
(to know where it was stored) but that is no longer required.