Rewrite how we search for libunwind (hopefully correctly this time)
This time, I used the canonical approach: Instead of passing the
absolute path to the libraries to the linker, I use link_directory()
and -lname as I should.

I hope that this version will work for both java on all our
target architectures, and also for stampede.

If so, we should maybe contribute it to libunwind and/or cmake projects.