[ENERGY] The energy now depends on the HostLoad plugin.
When a pstate is changed, the energy plugin will update a timestamp for that
host even if that host is never used; this means that we cannot rely on this
to determine if a host has been idle or not. The HostLoad plugin allows
us to do that.

This commit also updates several tesh files as some hosts were reported
as active in the past although they have not computed anything. These
hosts were hence idle and should be reported as such; a machine that
only boots or shuts down is in this case also regarded as unused.