use SPLITDUPLEX internally instead of FULLDUPLEX
We will s/FULLDUPLEX/SPLITDUPLEX/ everywhere in SimGrid. Rational:

That sharing mode is modeled by spliting every link between two links:
one uplink used when traversing the link in a direction, and the
downlink when we traverse the same link in the other direction.

With SHARED links, the bandwidth is shared between all flows without
taking the flow direction into account.

So, SPLITDUPLEX is the way we model fullduplex links, but at the end,
each of the SimGrid link in a SPLITDUPLEX (either the up or down link)
can be used in one direction only. That's the right model, but using
FULLDUPLEX for the situation where you have 2 links was rather
disturbing.

And yes, users will soon notice, when they will try to use network
energy models.