[tesh] kill process -> kill whole process group
If the command launched by tesh executes some subprocesses
(which are not cleaned by the root process on SIGKILL)
the previous code let such subprocesses in a running state.

Now the root process is launched into a new session (man setsid)
and the whole session is killed by tesh. All subprocess should be killed
as long as the session is not changed in the process tree.

Traversing the process tree could also be done, but it may require
additional dependencies (i.e., psutil).