[INSTR] Move vector* to shared_ptr<vector>.
By initializing an object, for instance of type VarCollTIData,
the vectors containing receive and send counts were deleted
when the TI object was deleted. This is unexpected by a 'normal'
user - by accepting shared_ptrs, this can be ameliorated.

The constructor that still accepts vector* should be removed once
all initializations use the constructor accepting the shared_ptr.