kill an unused function
The prototype was s::k::actor::set_maestro while its declaration was
simgrid::simix::set_maestro (since one commit), proving that it's not
used.

The used function is SIMIX_set_maestro(), which should be converted
to C++ for sure, but I'm not sure yet of how to do it properly.