[SMPI] Replay: Enable usage of tags
With this commit, tags of messages sent through the MPI Send/Recv
mechanism(s) are recorded in traces and also used when replayed.

This allows SimGrid to replay also more complex scenarios.

Tests and traces will be adjusted in a separate commit to prevent
this commit from being bloated. (This means, however, that the
smpi-replay tests will fail in this commit.)