have the maxmin system create by itself what it needs for selective update
Earlier, the modified_set was created externally by some models (but
not all). This lead to the interesting situation where we were asking
for selective update with the boolean field, but not getting this
because the modified_set was not existing.

Now that we get it when we ask for it, we detect that:
- the lmm_usage tests cannot activate selective_update because they
  don't have real actions, so they cannot track them to see which were
  modified.
- Storage is having selective updates, but with full algorithm (thus
  the change to Action::~Action() so that we deal when selective is on
  but not Lazy). No idea of whether this is the wanted behavior.