[SMPI] Add code to support std::tuple for std::unordered_map
This is currently in the replay file because it's only needed there.
I hope that a better solution can be proposed (or we should move
this to a header file).

The reason for adding this code here is that it will be used in
future commits (for identifying replay send/recv actions through
triplets, (src,dst,tag)).

This commit has been split from others to make the commits smaller.