[SMPI/TESH] Breaks tests: Fix colors for functions
The map with the colors used in instr_smpi.cpp was not correctly
checked because substrings were allowed too:

If irecv was passed, the map found recv first, checked whether
recv is a substring of irecv (= yes) and returned the color associated
with recv even though irecv was in the map as well.

The bugfix will be in the next commit.