SMPI does not need to have its own userdata pointer anymore
Before, the field of actors were used to store the SMPI data directly,
but now it's free.

With a bit of luck, this change will help StarPU
which fails since a few days, possibly because it tries to get the
userdata from an actor using smpi_process_get_userdata while that
actor is not from SMPI (I'm not sure, that's just a wild guess).

If my guess is right, retriving the userdata from the s4u::Actor will
certainly help. If it's wrong, the change shouldn't harm anything and
it will improve by a tiny bit our memory footprint.