changing the way the tracing category is passed to ExecImpl
the former version implied to set the category AFTER the start()
because the ExecImpl had to exist. It was against the idea of "do all
the sets and then starts". Now, the category is set at user level and
then passed to the Impl within the start simcall. Setting a category
after the start is now forbidden.