[HostLoad] Add method HostLoad::add_activity(ExecImplPtr)
The HostLoad plugin keeps track of all the computations that are going on
and does no longer have to rely on calculations of the sort time_passed*speed.

Instead, we can now access the total cost of the computation and keep track
of how much is still remaining. Like this, we always account for exactly
the cost of that calculation.