get rid od auto_restart_processes_
all actors are now put in actors_at_boot_
when a host is turned off, all the actors in actors_at_boot_ that do
not have the auto_restart_ flag to true are removed from the map. This
way, only the the actors that have to restart will when the host is
turned back on.