[DVFS] Support min/max pstate ranges
Through the newly introduced plugin/dvfs/min-pstate and
plugin/dvfs/max-pstate config options, it is now possible
to configure a range of pstates which can be used
for dvfs.

This makes it possible to define floprates for startup etc
even though these pstates will never be used during the execution
of a program.