Mark actors as auto_restart when recreating them on restart.
I guess it often works without it because we restart the actors that
are in the actors_at_boot_ vector, and that list is sufficiently
accurate.

But still, several locations of the code actually look if the actor is
marked auto_restart_ without checking the vector, so it's much better
having redundent information source to contain the same data.