docs: show the sections in manual's chapters
Trick ahead: the sections are regular chapters, but with UTF
non-breaking spaces before their names to visually indent them a bit
in the sidebar.

The main page should now be changed because it looks ugly with that
trick. But the content had to be changed anyway, maybe replace with
the current content of the main webpage, that in turn should be
emptied of all letters to put nice pictures instead