[ENERGY] Add getter for idle consumption
We currently still have one idle value per pstate, however this
should change soon. For this reason, this implementation
only uses the value from pstate 0 and will move towards the
next syntax.

Doing this allows us to avoid changing the signature of the
function (as the pstate will then be no longer required).