Add the Catch header
This is a header-only library for unit-testing. Including the header
in our project is the recommended way of using it. No need to test for
its presence, then.

Boost::unit_test_framework shall soon vanish from our tree. Testing
for its presence is too bothersome, and Catch seems easier to use
anyway.