Yet again change the way SMPI_SAMPLE_MACRO works There is no need anymore to have the body of the loop inside the macro. Kudos to Arnaud L for the ideas.
now usage is just turning
for(x;y;z) into SMPI_SAMPLE_GLOBAL(x,y,z,iters,thres)

Time injection will be done only once for all the section, speeding up the simulation.

TODO : iterator value will not be correct at the end of the loop for now.
Not important if iterator is local to the loop, but may be annoying if not.