Copy the values of {C,CXX,F}FLAGS into smpi{cc,cxx,ff,f90}.
The goal is to be able to give compiler flags at cmake time, and to have
them used later by the smpi script.
For example, it can be used to pass "-m64" and build a 64bit simgrid
when it's 32it by default (e.g. on Solaris).

The goal is, for example, to be able to set the variable