Rename Action::priority into Action::sharing_penalty in surf as it should be
This commit contains the following at the s4u -> surf boundary:
  -        .set_priority(1. / priority_)
  +        .set_sharing_penalty(1. / priority_)
The old version was just sick.

This should be further propagated down into LMM, where the
sharing_penalty is still named "sharing_weight". Not a helpful name
for the poor souls who didn't wrote an hdr on this topic :)