LMM: rename sharing_weight into sharing_penalty for sake of clarity
This is similar to what just happened in surf yesterday. It allows:
  -   cnst.usage_ += elem.consumption_weight / elem.variable->sharing_weight_;
  +   cnst.usage_ += elem.consumption_weight / elem.variable->sharing_penalty_;
which seems far more understandable to me.

I'm glad I finally got the will to do this boooooring cleanup, after
hours of being puzzled whether a sharing_weight is actually an
advantage or a drawback for the action in the LMM. The answer was "it
depends on whether you are in app, surf or LMM". Now, you know that
having a high priority is the opposite of having a high penalty, and
things should be less intricate.