Improve how we search for pybind11
Include the searching modules in our tree. I am not absolutely
convinced that this is the right approach (we are dupplicating their
code, we'll have to maintain it), but the error message is rather
brutal without it, and plus, the result of "pip install pybind11" is
not enough without it.