cmake: Try harder to detect pybind11 when it's there and survive when not
This removes some cmake files that are internal to pybind11, because
on ubuntu-lts, there were a mismatch where it was taking the
system pybind11Config but the version of pybind11Tools that we
provide. The (older) system file was using a cmake function that is
not provided anymore by our recent pybind11Tools.

After this change, our pybind11Config is the upstream-provided one to
the exception that it survives to the absence of pybind11Tools (that
we don't provide). So two cases remain:

- our modified pybind11Config finds a usable pybind11Tools. This must
  be a system-wide pybind11Tools, meaning that pybin11 is installed and
  usable.
- our modified pybind11Config does not find any pybind11Tools. Don't
  try to use pybind11, but keep compiling SimGrid.