sanitize our detection of pybind11
For some reason, our modified version of pybind11Config.cmake was not
included in the archive. We use it to ensure that the error message is
human friendly when compiling on a machine where pybind is not
installed.

For some other reason, FindPythonLibsNew.cmake was still in our git. I
tried to pack it in simgrid at some point to achieve the above
objective, but it proved to be a bad idea. I removed the other cmake
helping file of pybind11 but it seems I forgot this one. Fixed now.