embeed the javasphinx module
I don't feel very good embeeding a software that was deprecated by its
sole author, but I don't see no alternative module, and we need it to
compile the documentation. I would like to distribute the
documentation as a Debian package (or maybe re-add it to the
tarballs), and I don't see any other options to get it working.