docs: embeeding javasphinx is a bad idea: it has too many dependencies
The rational to embeed it was that it's not included in Debian and it
won't be included any time soon since its author orphaned it, while
I'd like to distribute the doc as Debian packages.

But it won't work: javalang is one of the dependencies of javasphinx
that is not packaged either (even if it seems actively maintained
upstream), and it seems to not be the last dependency of javasphinx.

So let's stop this ugly hack. Too bad for the Debian doc. I could try
to have a multi-source package on Debian with the precompiled doc as
one of them, but since it's not the prefered format for modification,
I guess it won't get accepted. Not 100% sure.