revalidate last smpi-io test
location of disk changed when moving from remote_io.xml to
hosts_with_disks.xml (located on bob while on carl before). The shared
file is now opened locally by bob (was opened remotely on carl before)
and the recessary data transfers for remote accesses change to.
If we switch carl and bob in hostfile_io we observe the same behavior
as before, only the hostnames differ.