Revert "Add '-O0' flag for: cmake -DCMAKE_BUILD_TYPE=Debug"
The SimGrid way to request non-optimized builds is
-Denable_compile_optimizations=OFF

I agree that this is not very user-friendly, and your commit looked
nice, but it actually broke the build as enable_debug was set to false
in release mode. As a result, -DNDEBUG was passed to the compilation,
removing all xbt_assert() and leading to many compilation issues since
our code does not compile without the asserts.

This reverts commit a4d4e2488e05031945fa3f1dd887e34c05cfa948.