remove a now useless conversion script
The C++ code will hint how to do the conversion. That may be time
consuming for someone wanting to convert a large file manually, but
I'm not sure of how long we'd need that script around. And I think
that nobody has such a long XML file (but the PR authors) so that
should be OK.