change handling of nested <prop> declarations.
Before: distinctions between zone props and other props with a boolean
and a single set of properties at a time.

Issue: As host can declare props and disks that in turn may declare
props, it enforces a declaration order: disks then props.

Now: stack a new property set when parsing a opening tag to store props
declared at this level. When closing a tag, we assign the set at the
top the stack where it has to and remove it from the stack.