Add autodoxy, a sphinx extension heavily inspired from autodoc_doxygen
This version is really really close to the now abandonned
https://github.com/rmcgibbo/sphinxcontrib-autodoc_doxygen
but I plan to improve it in the next commits.

The only difference to the original code is what is needed to make it
compile here (change imports, add a new setup.py, don't name the
template as '*.rst' since sphinx spits an error on them) plus the
rename autodoc_doxygen -> autodoxy.

Used version:
https://github.com/rmcgibbo/sphinxcontrib-autodoc_doxygen/commit/ad70f62805affdeb0e6cc638344c15a213394e0d